Factor out control sys IsActive tag to new file
