+++ src/ControlSystem/Tags/CMakeLists.txt
@@ -12,6 +12,7 @@ spectre_target_headers(
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
   HEADERS
   FunctionsOfTimeInitialize.hpp
+  IsActive.hpp
   MeasurementTimescales.hpp
   OptionTags.hpp
   )
